fix alur pendaftaran santri baru & lama

+++ application/controllers/modul_pendaftaran/pendaftaran.php
@@ -67,9 +67,29 @@ class Pendaftaran extends CI_Controller{
         }
     }
 
+    public function tambah_santri_lama()
+    {
+        if($this->session->id_user=="")
+        {
+            redirect(base_url('index.php/auth'), 'refresh');
+        }
+        else
+        {
+            $data['data_pendaftaran']=$this->pendaftaran_model->tampil_pendaftaran_list_lama();
+            $data['page']='modul_pendaftaran/pendaftaran/tambah_santri_lama_pendaftaran_modul_pendaftaran_view';
+            $this->load->view('modul_pendaftaran/dasbor_modul_pendaftaran_view', $data);
+        }
+    }
+
     public function get_pendaftaran_exist()
     {
         $data=$this->pendaftaran_model->get_pendaftaran_exist();
+		echo json_encode($data);
+    }
+
+    public function get_santri_exist()
+    {
+        $data=$this->pendaftaran_model->get_santri_exist();
 		echo json_encode($data);
     }
 
@@ -82,10 +102,121 @@ class Pendaftaran extends CI_Controller{
 		echo json_encode($data);
     }
 
+    public function aksi_tambah_pendaftaran_lama()
+    {
+        //insert pendaftaran
+        $id_pendaftaran=$this->input->post('id_pendaftaran');
+        $this->barcode($this->input->post('id_pendaftaran'));
+        $data=$this->pendaftaran_model->tambah_pendaftaran_lama();
+        
+        if($data)
+        {
+            //insert daftar_ulang
+            $du=array
+            (
+                'id_pendaftaran'=>$this->input->post('id_pendaftaran'),
+                'infaq_daftar_ulang'=>'santri_lama',
+                'status_daftar_ulang'=>'1',
+            );
+            $this->db->insert('daftar_ulang',$du);
+            if($this->db->affected_rows()>0)    
+             {
+                $data=TRUE;
+             }
+             else {
+                 $data= FALSE;
+             }
+
+             //insert mst_santri
+             $a=$this->db->query('select * from pendaftaran where id_pendaftaran="'.$id_pendaftaran.'"')->result_array();
+             $n_santri         = $a[0]['n_pendaftaran']; //nama lengkap santri
+             $nl_santri        = $a[0]['nl_pendaftaran']; // nama panggilan santri
+             $t_santri         = $a[0]['t_pendaftaran']; //tempat lahir santri
+             $tl_santri        = $a[0]['tl_pendaftaran']; //tanggal lahir santri
+             $alamat_santri    = $a[0]['alamat_pendaftaran']; // alamat_santri
+             $telp_santri      = $a[0]['telp_pendaftaran']; //no telp santri
+             $email_santri     = $a[0]['email_pendaftaran']; //email atau fb santri
+             $instansi_santri  = $a[0]['instansi_pendaftaran']; //instansi/sekolah asal santri
+            
+             $obj=array(
+                 'id_santri'=>$id_pendaftaran,
+                 'n_santri'=>$n_santri,
+                 'nl_santri'=>$nl_santri,
+                 't_santri'=>$t_santri,
+                 'tl_santri'=>$tl_santri,
+                 'alamat_santri'=>$alamat_santri,
+                 'telp_santri'=>$telp_santri,
+                 'email_santri'=>$email_santri,
+                 'instansi_santri'=>$instansi_santri,
+             );
+             $this->db->insert('mst_santri', $obj);
+             if($this->db->affected_rows()>0)    
+              {
+                 $data=TRUE;
+              }
+              else {
+                  $data= FALSE;
+              }
+
+              //insert kelas_asrama
+            $obj_kelas_asrm=array('id_santri'=>$id_pendaftaran,);
+             $kls_asrm=$this->db->insert('kelas_asrama', $obj_kelas_asrm);
+             if($this->db->affected_rows()>0)    
+             {
+                $data=TRUE;
+             }
+             else {
+                 $data= FALSE;
+             }
+        }
+        else
+        {
+            $data=false;
+        }
+		echo json_encode($data);
+    }
+
     public function aksi_edit_pendaftaran()
     {   
-        //$data['data_id']=$this->input->post('id_pendaftaran');
+        $id_pendaftaran=$this->input->post('id_pendaftaran');
         $data['data_esp']=$this->pendaftaran_model->edit_pendaftaran();
+        if($data)
+        {
+            $a=$this->query('select * from mst_santri where id_santri="'.$id_pendaftaran.'"')->num_rows();
+            if($a>0)
+            {
+                   //edit mst_santri
+                    $a=$this->db->query('select * from pendaftaran where id_pendaftaran="'.$id_pendaftaran.'"')->result_array();
+                    $n_santri         = $a[0]['n_pendaftaran']; //nama lengkap santri
+                    $nl_santri        = $a[0]['nl_pendaftaran']; // nama panggilan santri
+                    $t_santri         = $a[0]['t_pendaftaran']; //tempat lahir santri
+                    $tl_santri        = $a[0]['tl_pendaftaran']; //tanggal lahir santri
+                    $alamat_santri    = $a[0]['alamat_pendaftaran']; // alamat_santri
+                    $telp_santri      = $a[0]['telp_pendaftaran']; //no telp santri
+                    $email_santri     = $a[0]['email_pendaftaran']; //email atau fb santri
+                    $instansi_santri  = $a[0]['instansi_pendaftaran']; //instansi/sekolah asal santri
+                    
+                    $obj=array(
+                        'n_santri'=>$n_santri,
+                        'nl_santri'=>$nl_santri,
+                        't_santri'=>$t_santri,
+                        'tl_santri'=>$tl_santri,
+                        'alamat_santri'=>$alamat_santri,
+                        'telp_santri'=>$telp_santri,
+                        'email_santri'=>$email_santri,
+                        'instansi_santri'=>$instansi_santri,
+                    );
+                    $this->db->where('id_santri',$id_pendaftaran);
+                    $this->db->update('mst_santri', $obj);
+                    if($this->db->affected_rows()>0)    
+                    {
+                        $data=TRUE;
+                    }
+                    else {
+                        $data= FALSE;
+                    }
+            }
+        }
 		echo json_encode($data);
     }
 
@@ -99,28 +230,41 @@ class Pendaftaran extends CI_Controller{
             
             if($du[0]['ijazah_daftar_ulang']!="")
             {
-                unlink('./assets/upload/ijazah/'.$du[0]['ijazah_daftar_ulang']);
+                if(file_exists('./assets/upload/ijazah/'.$du[0]['ijazah_daftar_ulang']))
+                {
+                    unlink('./assets/upload/ijazah/'.$du[0]['ijazah_daftar_ulang']);
+                }
             }else
             {
                 $data['ijazah']='gagal hapus data';
             }
              if($du[0]['kk_daftar_ulang']!="")
             {
-                unlink('./assets/upload/kk/'.$du[0]['kk_daftar_ulang']);
+                if(file_exists('./assets/upload/kk/'.$du[0]['kk_daftar_ulang']))
+                {
+                    unlink('./assets/upload/kk/'.$du[0]['kk_daftar_ulang']);
+                }
+                
             }else
             {
                 $data['kk']='gagal hapus data';
             }
              if($du[0]['infaq_daftar_ulang']!="")
             {
-                unlink('./assets/upload/infaq/'.$du[0]['infaq_daftar_ulang']);
+                if(file_exists('./assets/upload/infaq/'.$du[0]['infaq_daftar_ulang']))
+                {
+                    unlink('./assets/upload/infaq/'.$du[0]['infaq_daftar_ulang']);
+                } 
             }else
             {
                 $data['infaq']='gagal hapus data';
             }
              if($du[0]['yatim_daftar_ulang']!="")
             {
-                unlink('./assets/upload/yatim/'.$du[0]['yatim_daftar_ulang']);
+                if(file_exists('./assets/upload/yatim/'.$du[0]['yatim_daftar_ulang']))
+                {
+                    unlink('./assets/upload/yatim/'.$du[0]['yatim_daftar_ulang']);
+                } 
             }
             else
             {
@@ -130,9 +274,13 @@ class Pendaftaran extends CI_Controller{
         }
         if(count($p))
         {
-            if(file_exists('./assets/upload/santri/'.$p[0]['foto_pendaftaran']))
+            if($p[0]['foto_pendaftaran']!="")
             {
-                unlink('./assets/upload/santri/'.$p[0]['foto_pendaftaran']);
+                if(file_exists('./assets/upload/santri/'.$p[0]['foto_pendaftaran']))
+                {
+                    unlink('./assets/upload/santri/'.$p[0]['foto_pendaftaran']);
+                }
+                
             }
             else
             {
@@ -188,6 +336,42 @@ class Pendaftaran extends CI_Controller{
 
     }
 
+    public function aksi_upload_foto_lama()
+    {
+        $id_pendaftaran= $this->input->post('id_pendaftaran');
+        
+        $config['upload_path']    = "./assets/upload/santri"; //path folder file upload
+        $config['allowed_types']  = 'jpg|jpeg|png'; //type file yang boleh di upload
+        $config['encrypt_name']   = TRUE; //enkripsi file name upload
+
+        $this->load->library('upload',$config); //call library upload
+        if($this->upload->do_upload("foto"))
+        {
+            $data = array('upload_data' => $this->upload->data()); //ambil file name yang diupload
+            
+            $nama_file  = $data['upload_data']['file_name'];
+            $data_ins=array(
+                'foto_pendaftaran'=>$nama_file,
+            );
+
+            $this->db->where('id_pendaftaran',$id_pendaftaran);
+            $this->db->update('pendaftaran', $data_ins);
+            if($this->db->affected_rows()>0)    
+             {
+                header('Location:'.base_url().'index.php/modul_pendaftaran/pendaftaran/tambah_santri_lama');
+             }
+             else 
+             {
+                 echo 'Data gagal tersimpan !';
+             }
+        }
+        else
+        {
+            echo 'Upload file gagal !';
+        }
+
+    }
+
     public function aksi_hapus_foto()
     {
         $id_pendaftaran=$this->uri->segment('4');
@@ -202,6 +386,33 @@ class Pendaftaran extends CI_Controller{
             if(unlink('./assets/upload/santri/'.$foto_pendaftaran))
             {
                 header('Location:'.base_url().'index.php/modul_pendaftaran/pendaftaran');
+            }
+            else
+            {
+                echo 'Gagal hapus foto';
+            }
+            
+        }
+        else 
+        {
+            echo 'Data gagal tersimpan !';
+        }
+    }
+
+    public function aksi_hapus_foto_lama()
+    {
+        $id_pendaftaran=$this->uri->segment('4');
+        $foto_pendaftaran=$this->uri->segment('5');
+        $data=array(
+            'foto_pendaftaran'=>'',
+        );
+        $this->db->where('id_pendaftaran',$id_pendaftaran);
+        $this->db->update('pendaftaran',$data);
+        if($this->db->affected_rows()>0)    
+        {
+            if(unlink('./assets/upload/santri/'.$foto_pendaftaran))
+            {
+                header('Location:'.base_url().'index.php/modul_pendaftaran/pendaftaran/tambah_santri_lama');
             }
             else
             {
